# docs/swagger.yaml
definitions:
  api.JsonObject:
    additionalProperties: true
    type: object
...
  errors.GenericError:
    properties:
      code:
        type: integer
      err: {}
    type: object
  model.Application:
    properties:
      amount:
        description: 数量
        type: integer
      created_at:
        type: string
      disk_space:
        description: 存储空间
        type: number
      email:
        description: 邮箱地址
        type: string
      id:
        type: integer
      ip:
        description: IP地址
        type: string
      ip_city:
        description: 城市
        type: string
      ip_country:
        description: 国家
        type: string
      node_type:
        description: 节点类型
        type: integer
      status:
        description: 状态
        type: integer
      updated_at:
        type: string
      upstream_bandwidth:
        description: 上行带宽
        type: number
      user_id:
        description: 用户ID
        type: string
    type: object
  model.CacheEvent:
    properties:
      block_size:
        description: 文件大小
        type: number
      blocks:
        description: blocks
        type: integer
      carfile_cid:
        description: CID
        type: string
      created_at:
        type: string
      device_id:
        description: 设备ID
        type: string
      id:
        type: integer
      max_created_time:
        type: string
      time:
        description: 时间
        type: string
      updated_at:
        type: string
    type: object
...
  model.RetrievalEvent:
    properties:
      blocks:
        description: blocks
        type: integer
      created_at:
        type: string
      device_id:
        description: 设备ID
        type: string
      id:
        type: integer
      time:
        description: 时间
        type: string
      updated_at:
        type: string
      upstream_bandwidth:
        description: 上传流量
        type: number
    type: object
  model.ValidationEvent:
    properties:
      blocks:
        description: blocks
        type: integer
      created_at:
        type: string
      device_id:
        description: 设备ID
        type: string
      duration:
        description: 完成时长
        type: integer
      id:
        type: integer
      status:
        description: 状态
        type: integer
      time:
        description: 时间
        type: string
      updated_at:
        type: string
      upstream_traffic:
        description: 上传流量
        type: number
      validator_id:
        description: 验证人
        type: string
    type: object
info:
  contact: {}
paths:
...
  /create_application:
    post:
      consumes:
      - application/json
      description: GetDeviceInfoHandler is a function to create an application for
        activation code
      parameters:
      - description: 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Application'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JsonObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 申请激活码
      tags:
      - 申请激活码
...
  /get_applications:
    get:
      consumes:
      - application/json
      description: GetApplicationsHandler is a function to get a slice of record(s)
        from application table in database
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      - description: 分页页码，默认 1
        in: query
        name: page
        type: integer
      - description: 分页页数 默认 50
        in: query
        name: page_size
        type: integer
      - description: 排序， Enums(ASC, DESC)
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: 排序字段
        in: query
        name: order_field
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                list:
                  items:
                    $ref: '#/definitions/model.Application'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取申请激活码记录
      tags:
      - 申请激活码
  /get_cache_list:
    get:
      consumes:
      - application/json
      description: GetDeviceInfoHandler is a handler to get a slice of record(s) from
        CacheEvent table in database
      parameters:
      - description: 设备ID
        in: query
        name: device_id
        required: true
        type: string
      - description: 分页页码，默认 1
        in: query
        name: page
        type: integer
      - description: 分页页数 默认 50
        in: query
        name: page_size
        type: integer
      - description: 排序， Enums(ASC, DESC)
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: 排序字段
        in: query
        name: order_field
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                list:
                  items:
                    $ref: '#/definitions/model.CacheEvent'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取设备的缓存列表
      tags:
      - 矿工页
...
  /get_retrieval_list:
    get:
      consumes:
      - application/json
      description: GetRetrievalListHandler is a handler to get a slice of record(s)
        from RetrievalEvent table in database
      parameters:
      - description: 设备ID
        in: query
        name: device_id
        required: true
        type: string
      - description: 分页页码，默认 1
        in: query
        name: page
        type: integer
      - description: 分页页数 默认 50
        in: query
        name: page_size
        type: integer
      - description: 排序， Enums(ASC, DESC)
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: 排序字段
        in: query
        name: order_field
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                list:
                  items:
                    $ref: '#/definitions/model.RetrievalEvent'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取检索数据列表
      tags:
      - 矿工页
...
  /get_validation_list:
    get:
      consumes:
      - application/json
      description: GetValidationListHandler is a handler to get a slice of record(s)
        from ValidationEvent table in database
      parameters:
      - description: 设备ID
        in: query
        name: device_id
        required: true
        type: string
      - description: 分页页码，默认 1
        in: query
        name: page
        type: integer
      - description: 分页页数 默认 50
        in: query
        name: page_size
        type: integer
      - description: 排序， Enums(ASC, DESC)
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: 排序字段
        in: query
        name: order_field
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                list:
                  items:
                    $ref: '#/definitions/model.ValidationEvent'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取验证数据列表
      tags:
      - 矿工页
swagger: "2.0"